♻️ refactor(sharedUI): extract the session callback from PlaybackService
